refactor(core): extract stream_audio handler into api/stream module
